+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Fixed issue with transaction toast title](https://github.com/multiversx/mx-sdk-dapp/pull/1555)
+
 ## [[5.1.1](https://github.com/multiversx/mx-sdk-dapp/pull/1553)] - 2025-08-12
 
 - [Fixed sdk-core peerDependency options](https://github.com/multiversx/mx-sdk-dapp/pull/1552)