+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - [Fixed issue with transaction toast title](https://github.com/multiversx/mx-sdk-dapp/pull/1555)
+- [Added pnpm](https://github.com/multiversx/mx-sdk-dapp/pull/1554)
 
 ## [[5.1.1](https://github.com/multiversx/mx-sdk-dapp/pull/1553)] - 2025-08-12