+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[v4.2.6](https://github.com/multiversx/mx-sdk-dapp/pull/1531)] - 2025-08-13
+
+- [Fixed 'Token not found' error is thrown during extension login](https://github.com/multiversx/mx-sdk-dapp/pull/1537)
+
 ## [[v4.2.5](https://github.com/multiversx/mx-sdk-dapp/pull/1531)] - 2025-07-25
 
 - [Unified the nativeAuth token logic for login hooks](https://github.com/multiversx/mx-sdk-dapp/pull/1530)